Added combo Q mana slider and Q after AA only
Those 2 options allow you to have more control over the Q usage in combo
mode to not waste mana to be able to use E on your target.
Also this adds a new menu option, which is default on, that you will use
Q only after and auto attack to ensure the hitchance of the spell, this
can be disabled within the menu.